requirements.txt
qrcode==7.4.2
pillow==10.3.0
tk==0.1.0